propose fix fees
